requirements.txt
streamlit==1.28.0
python-dotenv
openai